Create A. Equal or Not Equal.cpp
outputstandard output
You had n positive integers a1,a2,…,an arranged in a circle. For each pair of neighboring numbers (a1 and a2, a2 and a3, ..., an−1 and an, and an and a1), you wrote down: are the numbers in the pair equal or not.

Unfortunately, you've lost a piece of paper with the array a. Moreover, you are afraid that even information about equality of neighboring elements may be inconsistent. So, you are wondering: is there any array a which is consistent with information you have about equality or non-equality of corresponding pairs?