Create Chef and Pal.cpp
Chef just had a big argument with his best friend. He is quite upset and doesn't want anything to remind him of his pal.

Chef has to construct a binary string S of length N. He would like it to be as far from a palindrome as possible, and so he would like S to satisfy the following condition:

Let T be the reverse of S. Then, Si≠Ti for every 1≤i≤N.
Can you help Chef construct S? If there are multiple possible strings that satisfy Chef's condition, you may print any of them.

Note 1: A string S is said to be a binary string if every character of S is either '0' or '1'.

Note 2: Let S be a string of length N. String T is said to be its reverse if Ti=SN+1−i for every 1≤i≤N. For example, the reverse of 10010 is 01001 and the reverse of 0110 is 0110.