Create Weird Values.cpp
You are given an array A=[A1,A2,…,AN] containing N integers.

Consider any subarray [AL,AL+1,…,AR] of this subarray (where 1≤L≤R≤N). Ther weirdness of this subarray, denoted w(L,R), is defined to be the number of indices L≤i≤R such that Ai equals the number of occurrences of Ai in this subarray.

Formally, we define w(L,R) for 1≤L≤R≤N as follows:

For every integer x, let fL,R(x) denote the number of indices L≤j≤R such that Aj=x.
w(L,R) then equals the number of L≤i≤R such that Ai=fL,R(Ai)
For example, let A=[1,3,5,2,1,2,9,5]. Then,

w(1,8)=2, which can be computed as follows. We have
f1,8(1)=2
f1,8(2)=2
f1,8(3)=1
f1,8(5)=2
f1,8(9)=1
The only indices which satisfy f1,8(Ai)=Ai are i=4 and i=6, and hence w(1,8)=2.

w(3,5)=1 — the subarray is [5,2,1], and each of 1,2,5 occur exactly once in this subarray. Of these, only 1 equals the number of its occurrences in this subarray.
Given A, your task is to compute the sum of the weirdness of all subarrays of A, i.e, the value
∑L=1N∑R=LNw(L,R)