Create Marbles .cpp
In the fourth game, Sang-Woo and Ali are competing against each other. Sang-Woo is the more intelligent one between them, and he plans to win the game by cheating. Ali is your childhood friend and you want to help him win this game. The game goes as follows.

Both of them are given 2 strings — S and P — of equal length (say N). The task is to find the minimum number of operations required to convert the string P into S. Every character of both S and P is either a lowercase English alphabet or ?.

First, before performing any operation, you must choose a single lowercase character from the English alphabet, and replace every occurrence of ? with this character. Note that every ? present in either S or P is replaced by the exact same character. This does not count as an operation.

Once the above step is done, you are allowed to perform the following two operations on P any number of times:

Pick an index 1≤i≤N such that Pi is a vowel, and change Pi to any consonant.
Pick an index 1≤i≤N such that Pi is a consonant, and change Pi to any vowel.
Note: The vowels are the 5 characters {a,e,i,o,u} — the other 21 are consonants.

Find the minimum number of operations required to convert P into S.