Create Dalgona Treat.cpp
N  players advance to the 2nd round, the game is revealed to be "HoneyComb".

The players have to construct an array A. Each and every player is given "Ppopgi" (Dalgona candy), and the ith player has to carve a number Ai in their "Ppopgi".

The players will qualify if the array A satisfies the following rules:

The sum of the squares of all the numbers carved should be a perfect square. Formally, ∑i=1n(Ai)2 = K2 , where K is a natural number.
The number of distinct values in the array should not exceed min(10,N).
Gi-Hun wants everyone to qualify, so he asks you for help. Help him find an array A which satisfies these conditions.

If there are multiple arrays that satisfy the conditions, you may find any of them.

Note: N can be large enough that the array cannot be printed directly. Instead, describe the array in terms of its frequencies, i.e, print (value,frequency) pairs, where each pair suggests the number of times the value occurs in the array. If this is unclear, please look at the sample inputs and outputs.