Create B. And It's Non-Zero.cpp
You are given an array consisting of all integers from [l,r] inclusive. For example, if l=2 and r=5, the array would be [2,3,4,5]. What's the minimum number of elements you can delete to make the bitwise AND of the array non-zero?

A bitwise AND is a binary operation that takes two equal-length binary representations and performs the AND operation on each pair of the corresponding bits.

Input
The first line contains one integer t (1≤t≤104) — the number of test cases. Then t cases follow.

The first line of each test case contains two integers l and r (1≤l≤r≤2⋅105) — the description of the array.

Output
For each test case, output a single integer — the answer to the problem