Update and rename 1.cpp to A. Robot Clearner.cpp
A robot cleaner is placed on the floor of a rectangle room, surrounded by walls. The floor consists of n rows and m columns. The rows of the floor are numbered from 1 to n from top to bottom, and columns of the floor are numbered from 1 to m from left to right. The cell on the intersection of the r-th row and the c-th column is denoted as (r,c). The initial position of the robot is (rb,cb).

In one second, the robot moves by dr rows and dc columns, that is, after one second, the robot moves from the cell (r,c) to (r+dr,c+dc). Initially dr=1, dc=1. If there is a vertical wall (the left or the right walls) in the movement direction, dc is reflected before the movement, so the new value of dc is −dc. And if there is a horizontal wall (the upper or lower walls), dr is reflected before the movement, so the new value of dr is −dr.

Each second (including the moment before the robot starts moving), the robot cleans every cell lying in the same row or the same column as its position. There is only one dirty cell at (rd,cd). The job of the robot is to clean that dirty cell.
Given the floor size n and m, the robot's initial position (rb,cb) and the dirty cell's position (rd,cd), find the time for the robot to do its job.