Create Balanced Stone Heaps.cpp
outputstandard output
There are n heaps of stone. The i-th heap has hi stones. You want to change the number of stones in the heap by performing the following process once:

You go through the heaps from the 3-rd heap to the n-th heap, in this order.
Let i be the number of the current heap.
You can choose a number d (0≤3⋅d≤hi), move d stones from the i-th heap to the (i−1)-th heap, and 2⋅d stones from the i-th heap to the (i−2)-th heap.
So after that hi is decreased by 3⋅d, hi−1 is increased by d, and hi−2 is increased by 2⋅d.
You can choose different or same d for different operations. Some heaps may become empty, but they still count as heaps.
What is the maximum number of stones in the smallest heap after the process?

Input