Create Favourite String of Chef (Div 3).cpp
A string S is called Chef's favourite if every substring chef in S

must have a substring code before it.

You are given a string S
of size N that contains both code and chef as a substring. Please determine if S

is Chef's favourite.

Note: A substring is a continuous part of string which can be obtained by deleting some (may be none) character's from the beginning and some (may be none) character's from the end.