Create Count Number of Peaks.cpp
You are given an integer N where N≤10. Consider any array A with length N where each element can either be 0, 1, or 2, we define f(A) as the number of extrema in A. You need to find the sum of f(A) over all possible arrays A

.

Note:

    In an array A

, we consider Ai as an extremum if it is strictly greater than both of its neighbors (i.e. Ai>Ai−1 and Ai>Ai+1), or if it is strictly smaller than both of its neighbors (i.e. Ai<Ai−1 and Ai<Ai+1
). Note that first and last elements are not counted as extremum.
Extrema is the plural of extremum.