Create 1030A In Search of an Easy Problem.cpp
When preparing a tournament, Codeforces coordinators try treir best to make the first problem as easy as possible. This time the coordinator had chosen some problem and asked n

people about their opinions. Each person answered whether this problem is easy or hard.

If at least one of these n
people has answered that the problem is hard, the coordinator decides to change the problem. For the given responses, check if the problem is easy enough.