Create B. Integers Shop.cpp
The integers shop sells n segments. The i-th of them contains all integers from li to ri and costs ci

coins.

Tomorrow Vasya will go to this shop and will buy some segments there. He will get all integers that appear in at least one of bought segments. The total cost of the purchase is the sum of costs of all segments in it.

After shopping, Vasya will get some more integers as a gift. He will get integer x

as a gift if and only if all of the following conditions are satisfied:

    Vasya hasn't bought x

.
Vasya has bought integer l
that is less than x
.
Vasya has bought integer r
that is greater than x

    . 

Vasya can get integer x

as a gift only once so he won't have the same integers after receiving a gift.

For example, if Vasya buys segment [2,4]
for 20 coins and segment [7,8] for 22 coins, he spends 42 coins and receives integers 2,3,4,7,8 from these segments. He also gets integers 5 and 6

as a gift.

Due to the technical issues only the first s
segments (that is, segments [l1,r1],[l2,r2],…,[ls,rs]

) will be available tomorrow in the shop.

Vasya wants to get (to buy or to get as a gift) as many integers as possible. If he can do this in differents ways, he selects the cheapest of them.

For each s
from 1 to n, find how many coins will Vasya spend if only the first s segments will be available.