Create ChefWM.cpp
Chef recently created a new Tiling Window Manager called ChefWM. He wants to find out the maximum number of columns of windows he can fit on the screen and he needs your help.

Consider the screen to be an axis-aligned rectangle in the xy
-plane, with its bottom-left corner at (0,0) and top-right corner at (N,M) (that is, the screen has width N units and height M

units). ChefWM works as follows:

    First, the screen is divided into one or more columns of equal integer width.
    Then, each column is divided into two or more windows of equal integer height. Note that all windows within a column must have the same height, but windows in different columns may have different heights.
    Finally, Chef would like the following condition to hold:
        No two windows from different columns can have their horizontal boundary share the same y

-coordinate (apart from y=0 and y=M

        ). See the figures below for more details.

It is not allowed to leave any empty space on the screen, i.e, the division performed in the first two steps above must cover the entire screen.

Here is an example of a tiling that satisfies all the conditions, for N=8
and M=6

.

Below are two tilings that satisfy all conditions except the last. The one on the left has a horizontal edge at y=2
in both columns, while the one on the right has a horizontal edge at y=2 in column 1 and column 3

.

What is the maximum number of columns that ChefWM can fit on the screen while following the above conditions? Note that you do not need to maximize the number of windows, you only need to maximize the number of columns.

It may be possible that no columns can be created under these conditions, in that case, print 0
.