Create Mean Equals Median.cpp
Chef gets confused between mean and median very often, and as a result, he has developed a dislike of arrays whose mean and median are not equal.

Chef has an array A
of N

elements. He can perform the following operation on it:

    Pick an index 1≤i≤N

and increase Ai by 1

    .

He would like to obtain an array whose mean and median are equal. Determine the minimum number of operations required to achieve this.

Note: The median of an array A
of length N is defined as follows: Sort the array A

. Then,

    If N

is even, the median is the (N2)th
element
If N
is odd, the median is the (N+12)th

    element

For example, the median of the array [3,4,1,2]
is 2 and the median of the array [3,4,1] is 3.