Create 1325B CopyCopyCopyCopyCopy.cpp
Ehab has an array a of length n. He has just enough free time to make a new array consisting of n

copies of the old array, written back-to-back. What will be the length of the new array's longest increasing subsequence?

A sequence a
is a subsequence of an array b if a can be obtained from b

by deletion of several (possibly, zero or all) elements. The longest increasing subsequence of an array is the longest subsequence such that its elements are ordered in strictly increasing order.
Input

The first line contains an integer t

 — the number of test cases you need to solve. The description of the test cases follows.

The first line of each test case contains an integer n
(1≤n≤105) — the number of elements in the array a

The second line contains n
space-separated integers a1, a2, …, an (1≤ai≤109) — the elements of the array a
The sum of n
across the test cases doesn't exceed 105.