Create Retrieve back the Array.cpp
Dazzler had an array of N

distinct non-negative integers. Somehow he lost the array, but he knows the bitwise XOR of all the elements in the array. You have to help him to retrieve the array.

You are given two positive integers N
and X. Construct an array A of N

elements with the following conditions:

    For each i

(1≤i≤N), 0≤Ai≤5⋅105
.
All the elements in the array A
should be pairwise distinct, i.e, Ai≠Aj if i≠j
The bitwise XOR of all the N
elements in the array should be equal to X, i.e, A1⊕A2⊕…⊕AN=X, where ⊕

    denotes the bitwise XOR operation.

If there are multiple possible solutions, you may print any of them