Create B. Make AP.cpp
Polycarp has 3 positive integers a, b and c

. He can perform the following operation exactly once.

    Choose a positive integer m

and multiply exactly one of the integers a, b or c by m 
Can Polycarp make it so that after performing the operation, the sequence of three numbers a
, b, c (in this order) forms an arithmetic progression? Note that you cannot change the order of a, b and c
Formally, a sequence x1,x2,…,xn
is called an arithmetic progression (AP) if there exists a number d (called "common difference") such that xi+1=xi+d for all i from 1 to n−1. In this problem, n=3
For example, the following sequences are AP: [5,10,15]
, [3,2,1], [1,1,1], and [13,10,7]. The following sequences are not AP: [1,2,4], [0,1,0] and [1,3,2]
You need to answer t
independent test cases.