Create E. Masha-forgetful.cpp
Masha meets a new friend and learns his phone number — s. She wants to remember it as soon as possible. The phone number — is a string of length m that consists of digits from 0 to 9

. The phone number may start with 0.

Masha already knows n
phone numbers (all numbers have the same length m). It will be easier for her to remember a new number if the s is represented as segments of numbers she already knows. Each such segment must be of length at least 2

, otherwise there will be too many segments and Masha will get confused.

For example, Masha needs to remember the number: s=
'12345678' and she already knows n=4 numbers: '12340219', '20215601', '56782022', '12300678'. You can represent s as a 3 segment: '1234' of number one, '56' of number two, and '78' of number three. There are other ways to represent s

Masha asks you for help, she asks you to break the string s
into segments of length 2 or more of the numbers she already knows. If there are several possible answers, print any of them.